Split `krate` into MVC modules
There is still more work to do to extract model logic from some of the
controller endpoints, however I'm focusing on getting things moved
around in bulk first, and then focus on extracting logic when practical.
