+++ src/controllers/krate/mod.rs
@@ -1,0 +1,6 @@
+pub mod search;
+pub mod publish;
+pub mod owners;
+pub mod follow;
+pub mod downloads;
+pub mod metadata;
