+++ src/lib.rs
@@ -74,7 +74,6 @@ pub mod download;
 pub mod git;
 pub mod github;
 pub mod http;
-pub mod krate;
 pub mod owner;
 pub mod render;
 pub mod schema;
@@ -91,7 +90,6 @@ pub mod models;
 pub mod views;
 
 mod local_upload;
-mod with_count;
 
 /// Used for setting different values depending on whether the app is being run in production,
 /// in development, or for testing.
@@ -130,13 +128,22 @@ pub fn middleware(app: Arc<App>) -> MiddlewareBuilder {
     let mut api_router = RouteBuilder::new();
 
     // Route used by both `cargo search` and the frontend
-    api_router.get("/crates", C(krate::search::search));
+    api_router.get("/crates", C(controllers::krate::search::search));
 
     // Routes used by `cargo`
-    api_router.put("/crates/new", C(krate::publish::publish));
-    api_router.get("/crates/:crate_id/owners", C(krate::owners::owners));
-    api_router.put("/crates/:crate_id/owners", C(krate::owners::add_owners));
-    api_router.delete("/crates/:crate_id/owners", C(krate::owners::remove_owners));
+    api_router.put("/crates/new", C(controllers::krate::publish::publish));
+    api_router.get(
+        "/crates/:crate_id/owners",
+        C(controllers::krate::owners::owners),
+    );
+    api_router.put(
+        "/crates/:crate_id/owners",
+        C(controllers::krate::owners::add_owners),
+    );
+    api_router.delete(
+        "/crates/:crate_id/owners",
+        C(controllers::krate::owners::remove_owners),
+    );
     api_router.delete("/crates/:crate_id/:version/yank", C(version::yank::yank));
     api_router.put(
         "/crates/:crate_id/:version/unyank",
@@ -152,11 +159,11 @@ pub fn middleware(app: Arc<App>) -> MiddlewareBuilder {
     api_router.get("/versions/:version_id", C(version::deprecated::show));
 
     // Routes used by the frontend
-    api_router.get("/crates/:crate_id", C(krate::metadata::show));
+    api_router.get("/crates/:crate_id", C(controllers::krate::metadata::show));
     api_router.get("/crates/:crate_id/:version", C(version::deprecated::show));
     api_router.get(
         "/crates/:crate_id/:version/readme",
-        C(krate::metadata::readme),
+        C(controllers::krate::metadata::readme),
     );
     api_router.get(
         "/crates/:crate_id/:version/dependencies",
@@ -172,17 +179,35 @@ pub fn middleware(app: Arc<App>) -> MiddlewareBuilder {
     );
     api_router.get(
         "/crates/:crate_id/downloads",
-        C(krate::downloads::downloads),
+        C(controllers::krate::downloads::downloads),
     );
-    api_router.get("/crates/:crate_id/versions", C(krate::metadata::versions));
-    api_router.put("/crates/:crate_id/follow", C(krate::follow::follow));
-    api_router.delete("/crates/:crate_id/follow", C(krate::follow::unfollow));
-    api_router.get("/crates/:crate_id/following", C(krate::follow::following));
-    api_router.get("/crates/:crate_id/owner_team", C(krate::owners::owner_team));
-    api_router.get("/crates/:crate_id/owner_user", C(krate::owners::owner_user));
+    api_router.get(
+        "/crates/:crate_id/versions",
+        C(controllers::krate::metadata::versions),
+    );
+    api_router.put(
+        "/crates/:crate_id/follow",
+        C(controllers::krate::follow::follow),
+    );
+    api_router.delete(
+        "/crates/:crate_id/follow",
+        C(controllers::krate::follow::unfollow),
+    );
+    api_router.get(
+        "/crates/:crate_id/following",
+        C(controllers::krate::follow::following),
+    );
+    api_router.get(
+        "/crates/:crate_id/owner_team",
+        C(controllers::krate::owners::owner_team),
+    );
+    api_router.get(
+        "/crates/:crate_id/owner_user",
+        C(controllers::krate::owners::owner_user),
+    );
     api_router.get(
         "/crates/:crate_id/reverse_dependencies",
-        C(krate::metadata::reverse_dependencies),
+        C(controllers::krate::metadata::reverse_dependencies),
     );
     api_router.get("/keywords", C(controllers::keyword::index));
     api_router.get("/keywords/:keyword_id", C(controllers::keyword::show));
@@ -206,7 +231,7 @@ pub fn middleware(app: Arc<App>) -> MiddlewareBuilder {
         "/me/crate_owner_invitations/:crate_id",
         C(crate_owner_invitation::handle_invite),
     );
-    api_router.get("/summary", C(krate::metadata::summary));
+    api_router.get("/summary", C(controllers::krate::metadata::summary));
     api_router.put("/confirm/:email_token", C(user::confirm_user_email));
     api_router.put("/users/:user_id/resend", C(user::regenerate_token_and_send));
     api_router.get("/site_metadata", C(site_metadata::show_deployed_sha));
