+++ src/models/mod.rs
@@ -3,12 +3,17 @@ pub use self::category::{Category, CrateCategory, NewCategory};
 pub use crate_owner_invitation::NewCrateOwnerInvitation;
 pub use dependency::{Dependency, Kind, ReverseDependency};
 pub use download::VersionDownload;
+pub use self::follow::Follow;
 pub use self::keyword::{CrateKeyword, Keyword};
-pub use krate::{Crate, CrateDownload, Follow, NewCrate};
+pub use self::krate::{Crate, CrateDownload, NewCrate};
 pub use owner::{CrateOwner, NewTeam, Owner, OwnerKind, Rights, Team};
 pub use user::{Email, NewUser, User};
 pub use token::ApiToken;
 pub use version::{NewVersion, Version};
 
+pub mod helpers;
+
 mod category;
+mod follow;
 mod keyword;
+pub mod krate;
