+++ src/views/krate_publish.rs
@@ -7,7 +7,7 @@ use std::collections::HashMap;
 use serde::{de, Deserialize, Deserializer, Serialize, Serializer};
 use semver;
 
-use krate::MAX_NAME_LENGTH;
+use models::krate::MAX_NAME_LENGTH;
 
 use models::Keyword as CrateKeyword;
 use models::Kind as DependencyKind;
